用户管理模块优化

+++ src/views/user/index.vue
@@ -1,9 +1,16 @@
 <script setup lang="ts">
 import { reactive } from 'vue'
+import { ElMessage, ElMessageBox } from 'element-plus'
 import { User } from '../../types/user'
 import QuickCrud from '../../components/QuickCrud/index.vue'
-import { getUserList, addUser, updateUser, deleteUser } from '../../api/user'
-import { columns, formItems, searchFormItems } from './index'
+import {
+  getUserPageList,
+  getUserList,
+  addUser,
+  updateUser,
+  deleteUser,
+} from '../../api/user'
+import { columns, formItems, searchFormItems, formTitle, page } from './index'
 
 const dataList: User[] = reactive([])
 const searchForm = reactive({
@@ -18,7 +25,12 @@ const form: User = reactive({
   address: '',
   create_time: '',
 })
-const load = () => {
+const load = (parmas: any) => {
+  // getUserPageList(parmas).then((res: User[]) => {
+  //   // page.total=res.total
+  //   dataList.length = 0
+  //   dataList.push(...res)
+  // })
   getUserList().then((res: User[]) => {
     dataList.length = 0
     dataList.push(...res)
@@ -30,7 +42,6 @@ const handleSearch = () => {
 const handleReset = () => {
   console.log('handleReset!')
 }
-
 const handleAdd = () => {
   form.user_id = ''
   form.user_name = ''
@@ -45,9 +56,20 @@ const handleEdit = (item: User) => {
   form.create_time = item.create_time
   form.address = item.address
 }
+
 const handleDelete = (item: User, done: any) => {
-  deleteUser(item.user_id).then(() => {
-    done()
+  ElMessageBox.confirm(`你真的删除【${item.user_name}】的用户吗？`, '警告', {
+    confirmButtonText: '确定',
+    cancelButtonText: '取消',
+    type: 'warning',
+  }).then(() => {
+    deleteUser(item.user_id).then(() => {
+      ElMessage({
+        type: 'success',
+        message: '用户删除成功',
+      })
+      done()
+    })
   })
 }
 const handleCancel = () => {
@@ -56,10 +78,18 @@ const handleCancel = () => {
 const handleOk = (item: User, done: any) => {
   if (item.user_id) {
     updateUser(item).then(() => {
+      ElMessage({
+        type: 'success',
+        message: '用户修改成功',
+      })
       done()
     })
   } else {
     addUser(item).then(() => {
+      ElMessage({
+        type: 'success',
+        message: '用户创建成功',
+      })
       done()
     })
   }
@@ -79,6 +109,8 @@ const handleCurrentChange = (val: number) => {
     :form-items="formItems"
     :search-form-model="searchForm"
     :search-form-items="searchFormItems"
+    :form-title="formTitle"
+    :page="page"
     @on-load="load"
     @on-add="handleAdd"
     @on-edit="handleEdit"
